Move ReplaceStdlibShiftPass to separate file
