+++ src/opt.cpp
@@ -130,8 +130,6 @@ static llvm::Pass *CreateDebugPass(char *output);
 static llvm::Pass *CreateDebugPassFile(int number, llvm::StringRef name, std::string dir);
 #endif
 
-static llvm::Pass *CreateReplaceStdlibShiftPass();
-
 #ifdef ISPC_XE_ENABLED
 static llvm::Pass *CreateXeGatherCoalescingPass();
 static llvm::Pass *CreateReplaceLLVMIntrinsics();
@@ -665,102 +663,6 @@ static llvm::Pass *CreateDebugPassFile(int number, llvm::StringRef name, std::st
     return new DebugPassFile(number, name, dir);
 }
 #endif
-
-/** Given an llvm::Value known to be an integer, return its value as
-    an int64_t.
-*/
-static int64_t lGetIntValue(llvm::Value *offset) {
-    llvm::ConstantInt *intOffset = llvm::dyn_cast<llvm::ConstantInt>(offset);
-    Assert(intOffset && (intOffset->getBitWidth() == 32 || intOffset->getBitWidth() == 64));
-    return intOffset->getSExtValue();
-}
-
-///////////////////////////////////////////////////////////////////////////
-// ReplaceStdlibShiftPass
-
-class ReplaceStdlibShiftPass : public llvm::FunctionPass {
-  public:
-    static char ID;
-    ReplaceStdlibShiftPass() : FunctionPass(ID) {}
-
-    llvm::StringRef getPassName() const { return "Resolve \"replace extract insert chains\""; }
-
-    bool runOnBasicBlock(llvm::BasicBlock &BB);
-
-    bool runOnFunction(llvm::Function &F);
-};
-
-char ReplaceStdlibShiftPass::ID = 0;
-
-// This pass replaces shift() with ShuffleVector when the offset is a constant.
-// rotate() which is similar in functionality has a slightly different
-// implementation. This is due to LLVM(createInstructionCombiningPass)
-// optimizing rotate() implementation better when similar implementations
-// are used for both. This is a hack to produce similarly optimized code for
-// shift.
-bool ReplaceStdlibShiftPass::runOnBasicBlock(llvm::BasicBlock &bb) {
-    DEBUG_START_PASS("ReplaceStdlibShiftPass");
-    bool modifiedAny = false;
-
-    llvm::Function *shifts[6];
-    shifts[0] = m->module->getFunction("shift___vytuni");
-    shifts[1] = m->module->getFunction("shift___vysuni");
-    shifts[2] = m->module->getFunction("shift___vyiuni");
-    shifts[3] = m->module->getFunction("shift___vyIuni");
-    shifts[4] = m->module->getFunction("shift___vyfuni");
-    shifts[5] = m->module->getFunction("shift___vyduni");
-
-    for (llvm::BasicBlock::iterator iter = bb.begin(), e = bb.end(); iter != e; ++iter) {
-        llvm::Instruction *inst = &*iter;
-
-        if (llvm::CallInst *ci = llvm::dyn_cast<llvm::CallInst>(inst)) {
-            llvm::Function *func = ci->getCalledFunction();
-            for (int i = 0; i < 6; i++) {
-                if (shifts[i] && (shifts[i] == func)) {
-                    // we matched a call
-                    llvm::Value *shiftedVec = ci->getArgOperand(0);
-                    llvm::Value *shiftAmt = ci->getArgOperand(1);
-                    if (llvm::isa<llvm::Constant>(shiftAmt)) {
-                        int vectorWidth = g->target->getVectorWidth();
-                        int *shuffleVals = new int[vectorWidth];
-                        int shiftInt = lGetIntValue(shiftAmt);
-                        for (int i = 0; i < vectorWidth; i++) {
-                            int s = i + shiftInt;
-                            s = (s < 0) ? vectorWidth : s;
-                            s = (s >= vectorWidth) ? vectorWidth : s;
-                            shuffleVals[i] = s;
-                        }
-                        llvm::Value *shuffleIdxs = LLVMInt32Vector(shuffleVals);
-                        llvm::Value *zeroVec = llvm::ConstantAggregateZero::get(shiftedVec->getType());
-                        llvm::Value *shuffle =
-                            new llvm::ShuffleVectorInst(shiftedVec, zeroVec, shuffleIdxs, "vecShift", ci);
-                        ci->replaceAllUsesWith(shuffle);
-                        modifiedAny = true;
-                        delete[] shuffleVals;
-                    } else if (g->opt.level > 0) {
-                        PerformanceWarning(SourcePos(), "Stdlib shift() called without constant shift amount.");
-                    }
-                }
-            }
-        }
-    }
-
-    DEBUG_END_PASS("ReplaceStdlibShiftPass");
-
-    return modifiedAny;
-}
-
-bool ReplaceStdlibShiftPass::runOnFunction(llvm::Function &F) {
-
-    llvm::TimeTraceScope FuncScope("ReplaceStdlibShiftPass::runOnFunction", F.getName());
-    bool modifiedAny = false;
-    for (llvm::BasicBlock &BB : F) {
-        modifiedAny |= runOnBasicBlock(BB);
-    }
-    return modifiedAny;
-}
-
-static llvm::Pass *CreateReplaceStdlibShiftPass() { return new ReplaceStdlibShiftPass(); }
 
 #ifdef ISPC_XE_ENABLED
